backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
gunicorn==21.2.0
httpx==0.25.0
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
tenacity==8.2.3
python-dateutil==2.8.2